pyproject.toml
[tool.poetry]
name = "speech-anomaly"
version = "0.1.0"
description = "Voice emotion anomaly detection with scikit-learn models and XGBoost"
authors = ["Diego Incógnito <proton.me@diego.inc>"]
readme = "README.md"
# This replaces the 'where = ["src"]' part
packages = [
	{include = "batch", from = "src"},
	{include = "core", from = "src"},
	{include = "data", from = "src"},
	{include = "evaluation", from = "src"},
	{include = "experiments", from = "src"},
	{include = "features", from = "src"},
	{include = "models", from = "src"},
	{include = "utils", from = "src"},
]

[tool.poetry.dependencies]
python = ">=3.10"
hydra-core = ">=1.3.0"
omegaconf = ">=2.3.0"
pandas = ">=2.0.0"
numpy = ">=1.24.0"
librosa = ">=0.10.0"
scikit-learn = ">=1.3.0"
matplotlib = ">=3.7.0"
seaborn = ">=0.13.0"
wandb = ">=0.18.0"
umap-learn = ">=0.5.0"
pyarrow = "22.0.0"
praat-parselmouth = "0.4.7"
ffmpeg-python = "0.2.0"

[tool.poetry.group.bloat.dependencies]
xgboost = "3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"